Working on APIs - Exam.c now has comments for easier reading - base README given title and unhelpful info - Foundation/README.md created and gives the user some basic details on the structure and how to run the unit tests - Added unimplemented API for random, but it isn't done yet. I'm not quite happy with the design I ahve. - Added functions for min, max, and clamp with macros, but I have a comment which basically says I don't think that they are teh right approach in the long term, so I think I'll go the other way and implement individual functions - removed f_sqrt_f and f_sqrt_d, and updated linear_algebra.c to use c's sqrt and sqrtf instead.

+++ Exam/exam.c
@@ -1,15 +1,20 @@
-#include "exam.h"
 #include <stdio.h>
-#include <sys/types.h>
 
+#include "exam.h"
+#include "foundation.h"
+
+///////////////////////////////////////////////////////////////////////////////
+/// hidden globals
 char* _e_name;
 
-uint _e_passed = 0;
-uint _e_total  = 0;
+u64 _e_passed = 0;
+u64 _e_total  = 0;
 
-uint _e_all_passed = 0;
-uint _e_all_total = 0;
+u64 _e_all_passed = 0;
+u64 _e_all_total = 0;
 
+///////////////////////////////////////////////////////////////////////////////
+/// Function implementations
 void e_begin(char* name)
 {
     _e_name = name;
@@ -40,15 +45,15 @@ void e_end()
         ++_e_all_passed;
         printf("\x1b[1F"); // Move to beginning of previous line
         printf("\x1b[2K"); // Clear entire line
-        printf("%s Passed: %u / %u\n", _e_name, _e_passed, _e_total);
+        printf("%s Passed: %llu / %llu\n", _e_name, _e_passed, _e_total);
     }
     else
     {
-        printf("\tFailed: %u / %u\n", _e_passed, _e_total);
+        printf("\tFailed: %llu / %llu\n", _e_passed, _e_total);
     }
 }
 
 void e_log_summary()
 {
-    printf("\n%u/%u test cases pass.\n", _e_all_passed, _e_all_total);
+    printf("\n%llu/%llu test cases pass.\n", _e_all_passed, _e_all_total);
 }
